Added basic socket functionality for placing Placeable objects
Created PlacementPoint.cs that acts as a socket to place objects. Checks for incoming object ID so that only specific object can be placed. Checks for incoming object rotation, and checks it if is correct.
Created Placeable.cs - gets put on the grabbable object, contains the ID and boolean for if it's placed in a socket.
Demo in LT_TestScene.