Fix for placement point graphical glitches
Added extra checks for trigger enter/stay/exit - ignores player hands by tag, ignores the trigger enter and stay if it already has something snapped,  ignores trigger exit if the exiting object is not the object that was snapped. On object snapped, the placement points mesh renderer is disabled, and when the object exits the point, the renderer gets re-enabled.